修复realm数据库在组件化中与kotlin插件配置问题
