+++ app/src/main/java/com/ns/yc/lifehelper/ui/guide/contract/GuideContract.java
@@ -20,7 +20,6 @@ public interface GuideContract {
     }
 
     interface Presenter extends BasePresenter {
-        void cacheHomeNewsData();
         void cacheFindNewsData();
         void cacheFindBottomNewsData();
         void cacheHomePileData();
