+++ app/src/main/java/com/ns/yc/lifehelper/ui/main/contract/MainContract.java
@@ -27,7 +27,6 @@ public interface MainContract {
     interface Presenter extends BasePresenter {
         ArrayList<CustomTabEntity> getTabEntity();
         void getUpdate();
-        void locationPermissionsTask();
     }
