+++ app/src/main/java/com/ns/yc/lifehelper/ui/me/view/fragment/MeNewsCollectFragment.java
@@ -74,7 +74,6 @@ public class MeNewsCollectFragment extends BaseLazyFragment implements MeNewsCol
     @Override
     public void initView(View view) {
         recyclerView = view.findViewById(R.id.recyclerView);
-        presenter.initRealm();
         initRecycleView();
         initCallBack();
     }
